cmd_system/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.1; python_version >= '3.7'
cycler==0.11.0; python_version >= '3.6'
django-cors-headers==3.11.0
django==4.0.4
djangorestframework==3.13.1
fonttools==4.33.3; python_version >= '3.7'
kiwisolver==1.4.2; python_version >= '3.7'
matplotlib==3.5.2
numpy==1.22.3; python_version >= '3.8'
packaging==21.3; python_version >= '3.6'
pandas==1.4.2
pillow==9.1.0; python_version >= '3.7'
pykrige==1.6.1
pyparsing==3.0.8; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2022.1
scipy==1.8.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.4.2; python_version >= '3.5'
